backend/requirements.txt
flask-cors
dask
dask-image
scikit-learn
scikit-image
scipy
opencv-python
numpy
matplotlib
pandas
imutils